Refactor Output
